requirements.txt
absl-py==0.12.0
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
flatbuffers==1.12
gast==0.4.0
google-auth==1.28.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
gviz-api==1.9.0
h5py==3.1.0
idna==2.10
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tb-nightly==2.5.0a20210403
tensorboard-data-server==0.6.0
tensorboard-plugin-profile==2.4.0
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
tf-estimator-nightly==2.5.0.dev2021032601
tf-nightly==2.6.0.dev20210402
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
